+++ src/main/java/com/halisaha/backend/repository/SubPitchRepository.java
@@ -3,5 +3,5 @@ package com.halisaha.backend.repository;
 import com.halisaha.backend.model.SubPitch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubPitchRepository extends JpaRepository<SubPitch, Integer> {
+public interface SubPitchRepository extends JpaRepository<SubPitch, Long> {
 }
